@@ -9,16 +9,7 @@
     "@dnd-kit/sortable": "^7.0.1",
     "@dnd-kit/utilities": "^3.2.0",
     "@reduxjs/toolkit": "^1.9.0",
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.3.2",
-    "@testing-library/user-event": "^7.1.2",
-    "@types/jest": "^24.0.0",
-    "@types/node": "^12.0.0",
-    "@types/react": "^18.0.25",
-    "@types/react-dom": "^16.9.0",
-    "@types/react-image-gallery": "^1.0.4",
-    "@types/react-redux": "^7.1.7",
-    "@types/react-router-dom": "^5.1.7",
+    "redux-thunk": "^2.4.1",
     "antd": "^5.0.0",
     "axios": "^0.26.1",
     "classnames": "^2.3.1",
@@ -26,17 +17,15 @@
     "heic2any": "^0.0.3",
     "moment": "^2.29.4",
     "ramda": "^0.27.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
     "react-iframe": "^1.8.0",
     "react-image-gallery": "^1.2.5",
-    "react-redux": "^7.2.0",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
+    "react-redux": "^8.0.2",
+    "react-router-dom": "^5.3.4",
+    "react-scripts": "^5.0.1",
     "reselect": "^4.0.0",
-    "sass": "^1.37.5",
-    "sass-resources-loader": "^2.1.1",
-    "typescript": "^4.2.3"
+    "sass-resources-loader": "^2.1.1"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -62,8 +51,17 @@
     ]
   },
   "devDependencies": {
+    "@testing-library/dom": "^8.19.1",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^14.4.3",
+    "@types/node": "^17.0.42",
     "@types/prettier": "^2.2.3",
     "@types/ramda": "^0.27.40",
+    "@types/react": "^18.0.25",
+    "@types/react-dom": "^18.0.5",
+    "@types/react-image-gallery": "^1.0.4",
+    "@types/react-router-dom": "^5.1.7",
     "@typescript-eslint/eslint-plugin": "^4.19.0",
     "@typescript-eslint/parser": "^4.19.0",
     "eslint-config-airbnb-typescript": "^12.3.1",
@@ -77,6 +75,8 @@
     "gh-pages": "^3.2.3",
     "lint-staged": "^10.5.4",
     "prettier": "^2.2.1",
+    "sass": "^1.52.3",
+    "ts-jest": "^28.0.4",
     "webpack-bundle-analyzer": "^4.4.0"
   },
   "gitHooks": {
