@@ -1,6 +1,6 @@
 import React from 'react'
-import ReactDOM from 'react-dom'
 import { BrowserRouter } from 'react-router-dom'
+import { createRoot } from 'react-dom/client'
 
 import { Provider } from 'react-redux'
 
@@ -14,13 +14,16 @@ import './styles/index.scss'
 // eslint-disable-next-line functional/immutable-data
 Object.typedKeys = Object.keys
 
-ReactDOM.render(
+const container = document.querySelector('#root')
+// eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+const root = createRoot(container!)
+
+root.render(
   <Provider store={store}>
     <BrowserRouter>
       <App />
     </BrowserRouter>
-  </Provider>,
-  document.querySelector('#root')
+  </Provider>
 )
 
 // If you want your app to work offline and load faster, you can change
